[No-issue] Fix responsiveness for multi-viewport (#12)
* fix(mobile): resize book fonts and spacing for better mobile visibility

Read toggle padding was too close to delete book button. Added a thicker padding to decrease error
clicks.

* feat(book_card): add confirmation popup when deleting a book

This better prevents a user from removing their hard, entered book and not making it a pain to re-do
their progress.

* style(mobile): set up responsiveness for book addition form

Works with column grid breakpoints to offer a smooth experience on mobile and desktop as far as
usability goes by providing pleasant sizes to work with based on the viewport width.

* style(*): add icons for a bit of fancy styling

Icons added to the main functionality buttons. * Add a book, * Close form, * Add book to collection,
* Heading.